Refactoring ccmcc report into separate controller
