+++ app/views/reports/ccmcc_data.html.slim
@@ -1,5 +1,5 @@
 =render('shared/report_dates',
-  path: reports_ccmcc_data_path,
+  path: report_ccmcc_data_path,
   header: "CCMCC data extract",
   form_name: 'forms_finance_report',
   filters: false)
