+++ app/views/reports/index.html.slim
@@ -28,5 +28,5 @@ dl.govuk-summary-list
       dd.govuk-summary-list__value Extract raw data by date for Analytical Services
   - if policy(:report).ccmcc_data?
     .govuk-summary-list__row
-      dt.govuk-summary-list__key =link_to 'CCMCC data extract', reports_ccmcc_data_path
+      dt.govuk-summary-list__key =link_to 'CCMCC data extract', report_ccmcc_data_path
       dd.govuk-summary-list__value Extract CCMCC data by date for Analytical Services
