+++ config/routes.rb
@@ -14,8 +14,11 @@ Rails.application.routes.draw do
   get 'letter_templates' => 'reports#letters'
   get 'reports/raw_data' => 'reports#raw_data'
   put 'reports/raw_data' => 'reports#raw_data_export'
-  get 'reports/ccmcc_data' => 'reports#ccmcc_data'
-  put 'reports/ccmcc_data' => 'reports#ccmcc_data_export'
+
+  namespace :report do
+    get 'ccmcc_data' => 'ccmcc_data#show'
+    put 'ccmcc_data' => 'ccmcc_data#data_export'
+  end
 
   get '/applications/new' => 'applications/build#create'
   resources :applications, only: [] do
